edit int => float

+++ 05_calculator/calculator.py
@@ -14,7 +14,7 @@ def main():
 
     try:
         check_validate(argv)
-        result = calculate(int(argv[1]), int(argv[2]), argv[3])
+        result = calculate(float(argv[1]), float(argv[2]), argv[3])
     except NotMatchArgvError as err:
         print('[ERROR]:   引数の数が足りません。')
         print('[USAGE]:   python calculator.py [1つ目の数値][2つ目の数値][演算子(+-*/)]')
